Removed some layer violations from brave_rewards component
fix https://github.com/brave/brave-browser/issues/17482

This fixes part of violations from brave_rewards to resolve intermittent
build failure from CI. Other violations will be fixed as a next step.
