+++ chromium_src/chrome/browser/service_sandbox_type.h
@@ -27,18 +27,4 @@ content::GetServiceSandboxType<brave::mojom::ProfileImport>() {
 #include "brave/components/ipfs/service_sandbox_type.h"
 #endif
 
-#if !defined(OS_ANDROID)  // Android will use default, which is kUtility.
-namespace bat_ledger {
-namespace mojom {
-class BatLedgerService;
-}  // namespace mojom
-}  // namespace bat_ledger
-
-template <>
-inline sandbox::policy::SandboxType
-content::GetServiceSandboxType<bat_ledger::mojom::BatLedgerService>() {
-  return sandbox::policy::SandboxType::kNoSandbox;
-}
-#endif  // !defined(OS_ANDROID)
-
 #endif  // BRAVE_CHROMIUM_SRC_CHROME_BROWSER_SERVICE_SANDBOX_TYPE_H_
